Refactored the entities, implemented the absolute limit, daily limit and balance function, implemented the update limit for transactions, added the duplicateexception

+++ src/main/java/com/example/demo/dtos/AccountResponse.java
@@ -1,14 +1,16 @@
 package com.example.demo.dtos;
 
-import com.example.demo.models.Account;
-import com.example.demo.models.AccountType;
+import com.example.demo.entity.Account;
+import com.example.demo.common.enums.AccountType;
+
+import java.math.BigDecimal;
 
 public record AccountResponse(
         String iban,
         AccountType type,
-        double balance,
-        double absoluteLimit,
-        double dailyLimit,
+        BigDecimal balance,
+        BigDecimal absoluteLimit,
+        BigDecimal dailyLimit,
         boolean active,
         Long ownerId,
         String ownerName
