+++ src/main/java/com/example/demo/dtos/UpdateLimitsRequest.java
@@ -3,8 +3,10 @@ package com.example.demo.dtos;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record UpdateLimitsRequest(
-        @NotNull Double absoluteLimit,
-        @NotNull @PositiveOrZero Double dailyLimit
+        @NotNull BigDecimal absoluteLimit,
+        @NotNull @PositiveOrZero BigDecimal dailyLimit
 ) {
 }
